Client refactoring: replace Storage and WebSocketService with interfaces
