+++ library/src/main/java/com/getstream/sdk/chat/model/Channel.java
@@ -513,7 +513,7 @@ public class Channel {
                 Log.i(TAG, "channel query: merged watchers " + channel.getChannelState().getWatchers().size());
                 // offline storage
 
-                getClient().storage().insertMessagesForChannel(channel, response.body().getMessages());
+                getClient().getStorage().insertMessagesForChannel(channel, response.body().getMessages());
 
                 callback.onSuccess(response.body());
             }
@@ -841,7 +841,7 @@ public class Channel {
     public void handleChannelUpdated(Channel channel) {
         extraData = channel.extraData;
         updatedAt = channel.updatedAt;
-        getClient().storage().insertChannel(channel);
+        getClient().getStorage().insertChannel(channel);
     }
 
     public void handleWatcherStart(Event event) {
@@ -864,7 +864,7 @@ public class Channel {
         if (getLastMessageDate() != null && getLastMessageDate().before(message.getCreatedAt())) {
             setLastMessageDate(message.getCreatedAt());
         }
-        getClient().storage().insertMessageForChannel(this, message);
+        getClient().getStorage().insertMessageForChannel(this, message);
     }
 
     public void handleMessageUpdatedOrDeleted(Event event) {
@@ -883,7 +883,7 @@ public class Channel {
                 if (i == channelState.getMessages().size() - 1)
                     channelState.setLastMessage(message);
 
-                getClient().storage().insertMessageForChannel(this, message);
+                getClient().getStorage().insertMessageForChannel(this, message);
                 break;
             }
         }
