+++ library/src/main/java/com/getstream/sdk/chat/rest/EventHandler.java
@@ -2,6 +2,7 @@ package com.getstream.sdk.chat.rest;
 
 import android.os.Handler;
 import android.os.Message;
+
 import com.getstream.sdk.chat.model.Event;
 
 
@@ -15,6 +16,6 @@ public class EventHandler extends Handler {
     @Override
     public void handleMessage(Message msg) {
         super.handleMessage(msg);
-        webSocketService.webSocketListener.onWSEvent((Event) msg.obj);
+        webSocketService.getWebSocketListener().onWSEvent((Event) msg.obj);
     }
 }
