+++ library/src/main/java/com/getstream/sdk/chat/storage/OnQueryListener.java
@@ -1,0 +1,7 @@
+package com.getstream.sdk.chat.storage;
+
+public interface OnQueryListener<T> {
+    void onSuccess(T object);
+
+    void onFailure(Exception e);
+}
